requirements.txt
langchain>=0.2.1
langchain-openai>=0.1.8
langchain-community>=0.2.1
langchain-text-splitters>=0.2.0
chromadb>=0.5.3
duckdb>=0.10.3
openai>=1.30.3
requests>=2.31.0
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
pybgpstream>=2.0.0
py-radix>=0.10.0
ipaddress>=1.0.23 